requirements.txt
# For database migration testing
pyodbc==5.0.1
psycopg2-binary==2.9.9
urllib3
requests
